+++ src/background/seelen_weg/weg_items_impl.rs
@@ -2,24 +2,29 @@ use lazy_static::lazy_static;
 use parking_lot::Mutex;
 use seelen_core::{
     handlers::SeelenEvent,
-    state::{PinnedWegItemData, WegAppGroupItem, WegItem, WegItems},
+    state::{PinnedWegItemData, WegAppGroupItem, WegItem, WegItems, WegTemporalItemsVisibility},
 };
-use std::sync::Arc;
+use std::{collections::HashMap, sync::Arc};
 use tauri::Emitter;
 
 use crate::{
-    error_handler::Result, seelen::get_app_handle, state::application::FULL_STATE,
-    windows_api::window::Window,
+    error_handler::Result,
+    seelen::get_app_handle,
+    state::application::FULL_STATE,
+    windows_api::{window::Window, MonitorEnumerator},
 };
 
-use super::icon_extractor::{extract_and_save_icon_from_file, extract_and_save_icon_umid};
+use super::{
+    icon_extractor::{extract_and_save_icon_from_file, extract_and_save_icon_umid},
+    SeelenWeg,
+};
 
 lazy_static! {
     pub static ref WEG_ITEMS_IMPL: Arc<Mutex<WegItemsImpl>> =
         Arc::new(Mutex::new(WegItemsImpl::new()));
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct WegItemsImpl {
     items: WegItems,
 }
@@ -40,6 +45,18 @@ impl WegItemsImpl {
         }
     }
 
+    pub fn emit_to_webview(&self) -> Result<()> {
+        let handle = get_app_handle();
+        for (monitor_id, items) in self.get_filtered_by_monitor()? {
+            handle.emit_to(
+                SeelenWeg::get_label(&monitor_id),
+                SeelenEvent::WegInstanceChanged,
+                items,
+            )?;
+        }
+        Ok(())
+    }
+
     pub fn on_stored_changed(&mut self, stored: WegItems) -> Result<()> {
         let mut handles = vec![];
         for item in self.iter_all() {
@@ -53,7 +70,7 @@ impl WegItemsImpl {
         for handle in handles {
             self.add(&Window::from(handle))?;
         }
-        get_app_handle().emit(SeelenEvent::WegInstanceChanged, self.get())?;
+        self.emit_to_webview()?;
         Ok(())
     }
 
@@ -175,7 +192,63 @@ impl WegItemsImpl {
         self.items.sanitize();
     }
 
+    pub fn update_window(&mut self, window: &Window) {
+        let searching = window.address();
+        for item in self.iter_all_mut() {
+            if let WegItem::Pinned(data) | WegItem::Temporal(data) = item {
+                let maybe_window = data.windows.iter_mut().find(|w| w.handle == searching);
+                if let Some(app_window) = maybe_window {
+                    app_window.title = window.title();
+                    break;
+                }
+            }
+        }
+    }
+
+    fn filter_by_monitor(&mut self, monitor_id: &str) {
+        for item in self.iter_all_mut() {
+            match item {
+                WegItem::Pinned(data) | WegItem::Temporal(data) => {
+                    data.windows.retain(|w| {
+                        let window = Window::from(w.handle);
+                        window
+                            .monitor()
+                            .device_id()
+                            .is_ok_and(|id| id == monitor_id)
+                    });
+                }
+                _ => {}
+            }
+        }
+    }
+
     pub fn get(&self) -> WegItems {
         self.items.clone()
+    }
+
+    pub fn get_filtered_by_monitor(&self) -> Result<HashMap<String, WegItems>> {
+        let mut result = HashMap::new();
+        let state = FULL_STATE.load();
+
+        for monitor in MonitorEnumerator::get_all_v2()? {
+            let monitor_id = monitor.device_id()?;
+            if !state.is_weg_enabled_on_monitor(&monitor_id) {
+                continue;
+            }
+            let mode = state.get_weg_temporal_item_visibility(&monitor_id);
+            match mode {
+                WegTemporalItemsVisibility::All => {
+                    result.insert(monitor_id, self.items.clone());
+                }
+                WegTemporalItemsVisibility::OnMonitor => {
+                    let mut weg_items = self.clone();
+                    weg_items.filter_by_monitor(&monitor_id);
+                    weg_items.items.sanitize();
+                    result.insert(monitor_id, weg_items.items);
+                }
+            }
+        }
+
+        Ok(result)
     }
 }